move config-async to src/config/expo
